pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "initial_skeleton_dbt"
version = "1.0.0"
description = "This project contains initial skeleton code for a dbt project and some example code snippets to get started."
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.12,<3.13"


dependencies = [
    "agate==1.9.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "babel==2.17.0",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "daff==1.4.2",
    "db-dtypes==1.4.3",
    "dbt-adapters==1.16.0",
    "dbt-bigquery==1.9.2",
    "dbt-common==1.25.0",
    "dbt-core==1.10.2",
    "dbt-extractor==0.6.0",
    "dbt-protos==1.0.330",
    "dbt-semantic-interfaces==0.8.5",
    "deepdiff==7.0.1",
    "docstring_parser==0.16",
    "fastjsonschema==2.21.1",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "google-auth-oauthlib==1.2.2",
    "google-cloud-aiplatform==1.99.0",
    "google-cloud-bigquery==3.34.0",
    "google-cloud-core==2.4.3",
    "google-cloud-dataproc==5.20.0",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-storage==2.19.0",
    "google-crc32c==1.7.1",
    "google-genai==1.22.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.73.1",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "importlib_metadata==8.7.0",
    "isodate==0.6.1",
    "Jinja2==3.1.6",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter_core==5.8.1",
    "leather==0.4.0",
    "MarkupSafe==3.0.2",
    "mashumaro==3.14",
    "more-itertools==10.7.0",
    "msgpack==1.1.1",
    "nbformat==5.10.4",
    "networkx==3.5",
    "numpy==2.3.1",
    "oauthlib==3.3.1",
    "ordered-set==4.1.0",
    "packaging==25.0",
    "pandas==2.3.0",
    "pandas-gbq==0.29.1",
    "parsedatetime==2.6",
    "pathspec==0.12.1",
    "platformdirs==4.3.8",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyarrow==20.0.0",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "pydata-google-auth==1.9.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-slugify==8.0.4",
    "pytimeparse==1.1.8",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "setuptools==80.9.0",
    "shapely==2.1.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "snowplow-tracker==1.1.0",
    "sqlparse==0.5.3",
    "tenacity==8.5.0",
    "text-unidecode==1.3",
    "traitlets==5.14.3",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "websockets==15.0.1",
    "zipp==3.23.0",
]

[tool.licensecheck]
using = "PEP361"